Added cert to Firefox authorities | fixer.sh
As Firefox doesn't use system wide cert by default, we have to add our own manually. Added 'firefox_ssl' function to automate this with the install.

It will remove the policies.json file, if one already exists. May be worth changing if you have one existing, but for this use-case, we do not. This is to prevent people from overwriting policies if they run twice.

Thanks